separating DBconfig and applying priority mechanics
