+++ src/main/scala/ai/privado/languageEngine/java/tagger/PrivadoTagger.scala
@@ -33,16 +33,16 @@ import ai.privado.languageEngine.java.passes.read.{
   MessagingConsumerReadPass
 }
 import ai.privado.languageEngine.java.tagger.collection.{CollectionTagger, GrpcCollectionTagger, SOAPCollectionTagger}
+import ai.privado.languageEngine.java.tagger.config.JavaDBConfigTagger
 import ai.privado.languageEngine.java.tagger.sink.{InheritMethodTagger, JavaAPITagger, MessagingConsumerCustomTagger}
 import ai.privado.languageEngine.java.tagger.source.{IdentifierTagger, InSensitiveCallTagger}
 import ai.privado.tagger.PrivadoBaseTagger
 import ai.privado.tagger.collection.WebFormsCollectionTagger
-import ai.privado.tagger.config.DBConfigTagger
 import ai.privado.tagger.sink.RegularSinkTagger
 import ai.privado.tagger.source.{LiteralTagger, SqlQueryTagger}
 import io.shiftleft.codepropertygraph.generated.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes.Tag
-import io.shiftleft.semanticcpg.language._
+import io.shiftleft.semanticcpg.language.*
 import org.slf4j.LoggerFactory
 import overflowdb.traversal.Traversal
 
@@ -65,7 +65,7 @@ class PrivadoTagger(cpg: Cpg) extends PrivadoBaseTagger {
 
     new InSensitiveCallTagger(cpg, ruleCache, taggerCache).createAndApply()
 
-    new DBConfigTagger(cpg).createAndApply()
+    new JavaDBConfigTagger(cpg).createAndApply()
 
     new RegularSinkTagger(cpg, ruleCache).createAndApply()
